pyproject.toml
[project]
name = "xinyiclaw-py-web"
version = "0.1.0"
description = "XinyiClaw Web - AI Agent with MiniMax API and Web UI"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.10,<4.0",
    "croniter>=6.0.0",
    "python-dotenv>=1.2.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xinyiclaw"]

[tool.ruff]
line-length = 180
